Changed fibonacci to long long

+++ Homework_18/Homework_18.cpp
@@ -2,7 +2,7 @@
 #include <iostream>
 using namespace std;
 
-int fibonacci(int n)
+long long fibonacci(int n)
 {
     if (n <= 0) return 0;
     if (n == 1) return 1;
